debug_pp警告対策
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@8467 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/teraterm/debug_pp.cpp
@@ -59,7 +59,7 @@ void CDebugPropPage::OnInitDialog()
 {
 	// popup
 	SetCheck(IDC_DEBUG_POPUP_ENABLE, UnicodeDebugParam.CodePopupEnable);
-	for (int i = 0; i < _countof(key_list); i++) {
+	for (int i = 0; i < (int)_countof(key_list); i++) {
 		SendDlgItemMessage(IDC_DEBUG_POPUP_KEY1, CB_ADDSTRING, 0, (LPARAM)key_list[i].key_str);
 		SendDlgItemMessage(IDC_DEBUG_POPUP_KEY2, CB_ADDSTRING, 0, (LPARAM)key_list[i].key_str);
 		if (UnicodeDebugParam.CodePopupKey1 == key_list[i].key_code) {
